huy commit

+++ Baocao2/Controllers/RolesController.cs
@@ -11,47 +11,60 @@ namespace Baocao2.Controllers
     [Route("api/[controller]")]
     [ApiController]
     [Authorize]
-    public class RoleController : ControllerBase
+    public class RolesController : ControllerBase
     {
         private readonly RoleService _roleService;
-        
-        public RoleController(RoleService roleService)
+        private readonly RolePermissionService _rolePermissionService;
+        private readonly UserService _userService;
+
+        public RolesController(RoleService roleService , RolePermissionService rolePermissionService , UserService userService)
         {
             _roleService = roleService;
+            _rolePermissionService = rolePermissionService;
+            _userService = userService;
         }
 
-        [HttpGet("roles")]
+        [HttpGet]
         [ActionFilter(PERMISSION_FIX.Role_LIST , PERMISSION_FIX.User_ADD , PERMISSION_FIX.User_EDIT)]
         public ResultModel GetList()
         {
             return _roleService.GetList();
         }
 
-        [HttpGet("roles/{id}")]
+        [HttpGet("{id}")]
         [ActionFilter(PERMISSION_FIX.Role_VIEW)]
         public ResultModel GetById(Guid id)
         {
             return _roleService.GetById(id);
         }
 
-        [HttpPost("roles")]
+        [HttpPost]
         [ActionFilter(PERMISSION_FIX.Role_ADD)]
         public ResultModel Insert([FromBody] Role role)
         {
             return _roleService.Insert(role);
         }
 
-        [HttpPut("roles/{id}")]
+        [HttpPut("{id}")]
         [ActionFilter(PERMISSION_FIX.Role_EDIT)]
         public ResultModel Update(Guid id, [FromBody] Role role)
         {
             return _roleService.Update(id, role);
         }
 
-        [HttpDelete("roles/{id}")]
+        [HttpDelete("{id}")]
         [ActionFilter(PERMISSION_FIX.Role_DELETE)]
-        public ResultModel Delete(Guid id)
+        public ResultModel Delete(Guid id) // roleId
         {
+            if(_userService.isRoleUsed(id))
+            {
+                return new ResultModel { IsSuccess = false, Code = ResultModel.ResultCode.NotOK, Message = "Vai trò đang được sử dụng, không thể xóa", Id = null, Object = null };
+            }
+            var xoaRolePermissionsRes = _rolePermissionService.Delete(id);
+            if (!xoaRolePermissionsRes.IsSuccess)
+            {
+                return xoaRolePermissionsRes;
+            }
             return _roleService.Delete(id);
         }
     }
